Add ContractRootDirectory tests. Minor fix for throwing IOException instead   of IllegalArgumentException. Add config key for whether to enable deleting   buckets; while it can be handy at times to be able to actually delete root   (as has been the default in the past) as long as the right recursive-delete   and empty root constraints are present, HDFS and POSIX expect deleting root   recursively to only remove the contents and naturally to leave the filesystem   "intact" (not delete the bucket). Set the new default to disable bucket   deletion.
  In the inherited test, skip the two overly destructive tests for integration
  testing purposes (but keep them enabled in unittests);
  testRmEmptyRootDirNonRecursive and testRmRootRecursive do pass when run against
  a dedicated bucket, but for our continuous-test purposes won't play well
  with concurrency, sharded tests, etc.

	Change on 2016/10/04 by dhuo <dhuo@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=135181172

+++ gcsio/src/main/java/com/google/cloud/hadoop/gcsio/GoogleCloudStorageFileSystem.java
@@ -235,8 +235,10 @@ public class GoogleCloudStorageFileSystem {
 
     LOG.debug("create({})", path);
     Preconditions.checkNotNull(path);
-    Preconditions.checkArgument(!FileInfo.isDirectoryPath(path),
-        "Cannot create a file whose name looks like a directory.");
+    if (FileInfo.isDirectoryPath(path)) {
+      throw new IOException(String.format(
+          "Cannot create a file whose name looks like a directory. Got '%s'", path));
+    }
 
     // Check if a directory of that name exists.
     if (options.checkNoDirectoryConflict()) {
@@ -404,7 +406,12 @@ public class GoogleCloudStorageFileSystem {
         gcs.waitForBucketEmpty(resourceId.getBucketName());
         bucketsToDelete.add(resourceId.getBucketName());
       }
-      gcs.deleteBuckets(bucketsToDelete);
+      if (options.enableBucketDelete()) {
+        gcs.deleteBuckets(bucketsToDelete);
+      } else {
+        LOG.info("Skipping deletion of buckets because enableBucketDelete is false: {}",
+            bucketsToDelete);
+      }
     }
   }
